Move ActionEvent into events module
